+++ src/ir/CMakeLists.txt
@@ -6,6 +6,7 @@ set(ir_SOURCES
   effects.cpp
   eh-utils.cpp
   export-utils.cpp
+  inlining-utils.cpp
   intrinsics.cpp
   lubs.cpp
   memory-utils.cpp
